% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_geobs.R
\name{connect_geobs}
\alias{connect_geobs}
\title{Se connecter à Géobs}
\usage{
connect_geobs(url = "https://geobs.eaufrance.fr/geobs/login.htm", id, mdp)
}
\arguments{
\item{url}{Caractères. URL du site. Par défaut, "https://geobs.eaufrance.fr/geobs/login.htm".}

\item{id}{Caractères. Identifiant de connexion (en général l'adresse email).}

\item{mdp}{Caractère. Mot de passe.}
}
\value{
La connexion de la session
}
\description{
Se connecter à Géobs
}
\examples{
\dontrun{
connect_geobs(
id = 'john.doe@ofb.gouv.fr',
mdp = 'mon_mot_de_passe'
)
}
}
